renamed indexer to reader. this is what this module does. In reading the indexer spec I realized I was acutually developing the reader. Indexing is automatically handled by the downloader.
