+++ unit_tests/reader/data/tsFiles/ts/txt/2/obs/fr/resources.json
